# static/swagger.yaml
swagger: '2.0'
info:
  title: "API name"
  description: "Brief description"
  version: "1.0.0"
host: "127.0.0.1:5000" #working on local host
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /customers:
    get:
      tags:
        - Customers
      summary: "Returns all customers"
      description: "Returns all customer's information (requires token)"
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Retrieve customers successfully"
          schema:
            $ref: "#/definitions/AllCustomers"
  /customeraccnt:
    get:
      tags:
        - CustomerAccnt
      summary: "Returns all customeraccnts"
      description: "Returns all customer accounts information"
      responses:
        200:
          description: "Retrieve customeraccnt successfully"
          schema:
            $ref: "#/definitions/AllCustomerAccounts"
        400:
          description: "Customer Accounts not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ErrorResponse:
    type: "object"
    properties:
      message:
        type: "string"
        description: "Error Message"
      status:
        type: "string"
        description: "Error status"
  AllCustomers:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Customer's ID"
      name:
        type: "string"
        description: "Customer's name"
      email:
        type: "string"
        description: "Customer's email"
      phone:
        type: "string"
        description: "Customer's phone"
      username: 
        type: "string"
        description: "Customer's username"
      role_id:
        type: "integer"
        description: "Customer's role ID"
...
  AllCustomerAccounts:
    type: "object"
    properties :
      username:
        type: "string"
        description: "Username associated with the account"
      password:
        type: "string"
        format: "float"
        description: "Password associated with the account"
      customer_id:
        type: "number"
        description: "Customer Id associated with the customer account"      
...